+++ src/main/sdl/renderbase.hpp
@@ -19,7 +19,7 @@ public:
     virtual void disable()                    = 0;
     virtual bool start_frame()                = 0;
     virtual bool finalize_frame()             = 0;
-    virtual void draw_frame(uint32_t* pixels) = 0;
+    virtual void draw_frame(uint16_t* pixels) = 0;
     void convert_palette(uint32_t adr, uint32_t r, uint32_t g, uint32_t b);
 
 protected:
